new function and  reservation
In the projectlist, the user can see which project is refused or
published. The user can it see it from the background color in the
projectlist.

Here it has also two text boxes for the reservation. The second textbox
will appear, when the user chose in the priority section "2er Team" or
"1er oder 2er Team"

I also add the Passwort forgot in the login site, but it's still in
progres.